pyproject.toml
[tool.poetry]
name = "inventory-vanity"
version = "0.1.0"
description = ""
authors = ["Yasser Azan"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
djangorestframework = "^3.15.2"
python-decouple = "^3.8"
drf-yasg = "^1.21.7"
django-cors-headers = "^4.4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
